# improve agg_list performance of chunked numerical data