Delaying a full rechunk can improve performance of many query, but some are slower.

This improves the following query by `~3x`.

```rust
fn main() -> Result<()> {
    let q = LazyCsvReader::new("/home/ritchie46/code/db-benchmark/data/G1_1e7_1e2_0_0.csv".into())
        .with_rechunk(false)
        .finish()?;

    let out = q
        .groupby([col("id2"), col("id4")])
        .agg([(pearson_corr(col("v1"), col("v2")).pow(2.0)).alias("r2")])
        .collect()?;
    Ok(())
}
```

# before

```
          6.685,79 msec task-clock                #    2,124 CPUs utilized          
             4.747      context-switches          #  710,013 /sec                   
               218      cpu-migrations            #   32,606 /sec                   
           173.797      page-faults               #   25,995 K/sec                  
    21.435.123.551      cycles                    #    3,206 GHz                    
    38.153.190.822      instructions              #    1,78  insn per cycle         
     5.010.104.768      branches                  #  749,366 M/sec                  
        24.141.092      branch-misses             #    0,48% of all branches        

       3,148083894 seconds time elapsed

       6,231878000 seconds user
       0,469579000 seconds sys
```

# after

```
          5.024,99 msec task-clock                #    4,879 CPUs utilized          
             4.305      context-switches          #  856,718 /sec                   
               118      cpu-migrations            #   23,483 /sec                   
           212.864      page-faults               #   42,361 K/sec                  
    14.441.547.675      cycles                    #    2,874 GHz                    
    21.711.917.241      instructions              #    1,50  insn per cycle         
     4.031.158.541      branches                  #  802,222 M/sec                  
        21.774.943      branch-misses             #    0,54% of all branches        

       1,029876114 seconds time elapsed

       4,531616000 seconds user
       0,502194000 seconds sys
```